Split system error handling to a separate class
