+++ hummingbird-client/src/main/java/com/vaadin/client/ApplicationConnection.java
@@ -24,7 +24,6 @@ import com.google.web.bindery.event.shared.Event;
 import com.google.web.bindery.event.shared.EventBus;
 import com.google.web.bindery.event.shared.HandlerRegistration;
 import com.vaadin.client.ApplicationConnection.ApplicationStoppedEvent;
-import com.vaadin.client.bootstrap.ErrorMessage;
 import com.vaadin.client.communication.MessageHandler;
 import com.vaadin.client.gwt.com.google.web.bindery.event.shared.SimpleEventBus;
 import com.vaadin.client.hummingbird.BasicElementBinder;
@@ -341,55 +340,6 @@ public class ApplicationConnection {
     }
 
     /**
-     * Shows the communication error notification.
-     *
-     * @param details
-     *            Optional details.
-     * @param statusCode
-     *            The status code returned for the request
-     *
-     */
-    public void showCommunicationError(String details, int statusCode) {
-        Console.error("Communication error: " + details);
-        showError(details, configuration.getCommunicationError());
-    }
-
-    /**
-     * Shows the authentication error notification.
-     *
-     * @param details
-     *            Optional details.
-     */
-    public void showAuthenticationError(String details) {
-        Console.error("Authentication error: " + details);
-        showError(details, configuration.getAuthorizationError());
-    }
-
-    /**
-     * Shows the session expiration notification.
-     *
-     * @param details
-     *            Optional details.
-     */
-    public void showSessionExpiredError(String details) {
-        Console.error("Session expired: " + details);
-        showError(details, configuration.getSessionExpiredError());
-    }
-
-    /**
-     * Shows an error notification.
-     *
-     * @param details
-     *            Optional details.
-     * @param message
-     *            An ErrorMessage describing the error.
-     */
-    protected void showError(String details, ErrorMessage message) {
-        showError(message.getCaption(), message.getMessage(), details,
-                message.getUrl());
-    }
-
-    /**
      * Checks if deferred commands are (potentially) still being executed as a
      * result of an update from the server. Returns true if a deferred command
      * might still be executing, false otherwise. This will not work correctly
@@ -490,15 +440,6 @@ public class ApplicationConnection {
     public ReconnectDialogConfigurationState getReconnectDialogConfiguration() {
         // FIXME from the server
         return new ReconnectDialogConfigurationState();
-    }
-
-    public void showError(String caption, String message, String details,
-            String url) {
-        // FIXME Not like this
-        Browser.getWindow().alert(caption + "\n" + message + "\n" + details);
-        if (url != null) {
-            WidgetUtil.redirect(url);
-        }
     }
 
 }
